move join related code into join_utils.rs
