+++ datafusion/src/physical_plan/mod.rs
@@ -622,6 +622,7 @@ pub mod functions;
 pub mod hash_aggregate;
 pub mod hash_join;
 pub mod hash_utils;
+pub mod join_utils;
 pub mod json;
 pub mod limit;
 pub mod math_expressions;
